+++ Nix/utilities.nix
@@ -1,0 +1,40 @@
+####### Utilities ##############################################################
+##                                                                            ##
+## * Install a list of command-line utilities                                 ##
+##                                                                            ##
+################################################################################
+
+{ config, pkgs, ... }:
+
+{
+  environment.systemPackages = with pkgs; [
+    bat
+    curl
+    dcfldd
+    emv
+    git
+    git-lfs
+    gnupg
+    htop
+    iftop
+    imagemagick
+    killall
+    lsof
+    mkpasswd
+    mosh
+    openssh
+    p7zip
+    ranger
+    rsync
+    screenfetch
+    sshfs
+    testdisk
+    trash-cli
+    tree
+    unzip
+    watch
+    wget
+    xz
+    zip
+  ];
+}